Refactor raw-editor
