+++ addon/commands/text-properties/remove-bold-command.ts
@@ -1,11 +1,11 @@
 import RemovePropertyCommand from "./remove-property-command";
 import boldProperty from "../../utils/rdfa/bold-property";
-import { RawEditor } from "../../editor/raw-editor";
+import Model from "@lblod/ember-rdfa-editor/model/model";
 
 export default class RemoveBoldCommand extends RemovePropertyCommand {
   name = "remove-bold"
-  constructor(editor: RawEditor) {
-    super(editor, boldProperty);
+  constructor(model: Model) {
+    super(model, boldProperty);
   }
 
 }
