+++ addon/commands/text-properties/set-property-command.ts
@@ -1,12 +1,12 @@
 import EditorProperty from "../../utils/ce/editor-property";
 import Command from "../command";
-import { RawEditor } from "../../editor/raw-editor";
 import { applyProperty } from "../../utils/ce/property-helpers";
+import Model from "@lblod/ember-rdfa-editor/model/model";
 
 export default abstract class SetPropertyCommand extends Command {
   protected property: EditorProperty;
-  constructor(editor: RawEditor, property: EditorProperty) {
-    super(editor);
+  constructor(model: Model, property: EditorProperty) {
+    super(model);
     this.property = property;
   }
   execute() {
